feat: instrument navigation operations (#36)
Updates the root span name and http.route attribute to contain the route that was navigated to, rather than the route that started the navigation. Also adds a navigation span with more details about the navigation operation.
Instead of instrumenting separate classes / methods that trigger navigation, this change uses AbstractNavigationStateRenderer which seems to be called in all navigation scenarios (page load, client-side, server-side, forward).