feat: instrument HeartbeatHandler (#56)
Add instrumentation for `HeartbeatHandler`, which adds a span, and updates to root span to mark it as heartbeat request.

I removed the `UIInstrumentation` as part of this. It was overwriting the root span name, and it seems with all the other instrumentations updating the root span name to something more specific it is not needed anymore. 

Closes #5