feat: configure trace level (#71)
Allows to configure a global trace level option that instrumentations can check for whether they should add some detail to a trace or not. The option supports three levels `minimum`, `default`, `maximum`. The default is `default`. The option can be changed using any of the configuration mechanism supported by OpenTelemetry, for example environment variables or system properties. For example, to change the trace level with a system property, add the following argument to the java process: `-Dotel.instrumentation.vaadin.trace-level=maximum`.

This PR is only intended to add the basic option mechanism, and adapts the `UidlRequestHandlerInstrumentation` to demonstrate its use. The remaining instrumentations will be changed in a separate PR.

Part of #62