refactor: mark root span as server span (#89)
Changes the `VaadinServiceInstrumentation` to mark the root spans it creates as server spans instead of internal ones, and marks the root span as having an error if a nested exception occurs. This is necessary to have several vendors (New Relic, DataDog, Dynatrace) include the spans in their error monitoring.
Add several HTTP semantic attributes that we lost from disabling the default HTTP request instrumentations.