refactor: instrument VaadinServlet for root span creation (#125)
## Description

Moves the root / server span creation from
`VaadinServiceInstrumentation` to `VaadinServletInstrumentation`, and
only creates a root span if it doesn't exist yet. This ensures that
there is always only one server span, even when enabling the default
servlet / application server instrumentations, and also ensures that
there is a server span for Vaadin instrumentations that are outside of
`VaadinService.handleRequest` calls, such as serving static files.