test: prevent jetty start when skipping tests
When skipping tests with '-DskipTests', the goals bound to pre and post
integration-test phase are still executed. Since tests are not executed,
it may happen that when the post-integration-test phase is reached,
the forked Jetty server is not yet started, so it will not be stopped.
This change moved the integration test configuration to a maven profile
activate only when the 'skipTests' property is not set.