test: reduced verbosity of mock server
Mock server produces a huge amount of log messages that are not useful
in the context of observability tests.
This change adds a system parameter to the maven failsafe plugin configuration
to set the log level for mock server to WARN, in order to get only severe
messages.
The log level can be changed setting the 'mockserver.logLevel' system property
in the maven command line